… go/vt/sqlparser/ast_rewriting_test.go
… go/vt/sqlparser/normalizer_test.go
… go/vt/vtgate/executor_test.go
… go/vt/vtgate/planbuilder/simplifier_test.go
… go/vt/vtgate/semantics/typer_test.go