## Description
Merges `RewriteAST` and `Normalizer` so both actions can be done together. The less we need to walk the AST, the faster we can work with queries.

```
                                  │ ../benchmarks/v0.txt │       ../benchmarks/v1.txt        │
                                  │        sec/op        │   sec/op     vs base              │
Planner/from_cases.json-gen4-20              7.416m ± 1%   7.143m ± 4%  -3.68% (p=0.004 n=6)
Planner/filter_cases.json-gen4-20            149.9m ± 0%   146.8m ± 0%  -2.04% (p=0.002 n=6)
Planner/large_cases.json-gen4-20             463.3µ ± 1%   461.6µ ± 1%       ~ (p=0.093 n=6)
Planner/aggr_cases.json-gen4-20              13.17m ± 1%   12.78m ± 1%  -2.96% (p=0.002 n=6)
Planner/select_cases.json-gen4-20            10.78m ± 1%   10.45m ± 2%  -3.09% (p=0.002 n=6)
Planner/union_cases.json-gen4-20             3.856m ± 0%   3.707m ± 1%  -3.87% (p=0.002 n=6)
geomean                                      8.098m        7.882m       -2.67%

                                  │ ../benchmarks/v0.txt │        ../benchmarks/v1.txt        │
                                  │         B/op         │     B/op      vs base              │
Planner/from_cases.json-gen4-20             4.566Mi ± 0%   4.405Mi ± 0%  -3.53% (p=0.002 n=6)
Planner/filter_cases.json-gen4-20           20.13Mi ± 0%   19.37Mi ± 0%  -3.77% (p=0.002 n=6)
Planner/large_cases.json-gen4-20            276.6Ki ± 0%   273.9Ki ± 0%  -0.97% (p=0.002 n=6)
Planner/aggr_cases.json-gen4-20             7.350Mi ± 0%   7.122Mi ± 0%  -3.09% (p=0.002 n=6)
Planner/select_cases.json-gen4-20           6.334Mi ± 0%   6.141Mi ± 0%  -3.04% (p=0.002 n=6)
Planner/union_cases.json-gen4-20            2.269Mi ± 0%   2.185Mi ± 0%  -3.69% (p=0.002 n=6)
geomean                                     3.713Mi        3.601Mi       -3.02%

                                  │ ../benchmarks/v0.txt │       ../benchmarks/v1.txt        │
                                  │      allocs/op       │  allocs/op   vs base              │
Planner/from_cases.json-gen4-20              107.9k ± 0%   101.9k ± 0%  -5.54% (p=0.002 n=6)
Planner/filter_cases.json-gen4-20            534.4k ± 0%   508.9k ± 0%  -4.77% (p=0.002 n=6)
Planner/large_cases.json-gen4-20             10.71k ± 0%   10.61k ± 0%  -0.92% (p=0.002 n=6)
Planner/aggr_cases.json-gen4-20              168.1k ± 0%   160.1k ± 0%  -4.77% (p=0.002 n=6)
Planner/select_cases.json-gen4-20            147.8k ± 0%   140.8k ± 0%  -4.77% (p=0.002 n=6)
Planner/union_cases.json-gen4-20             53.22k ± 0%   50.46k ± 0%  -5.20% (p=0.002 n=6)
geomean                                      96.68k        92.48k       -4.34%

```

## Related Issue(s)


## Checklist

-   [x] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [x] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required